Clean up decoder feature initialization and declaration.
Change how we list and initialize decoder features, so there is only one
ordered list of them, and all codes that eneds to know that order depends
on that single declaration, thus removing the risk of inconsistencies, such
as we have had at various times.

Remove the support for "none" as a segmentation model - still accepted with a
warning, but removed everywhere I could.