Better handling of non-breaking and optional hyphens.
tmx2lfl.pl and ce_tmx.pl now replace \_ (non-breaking hyphen in RTF and Trados)
by -, and strip out \- (optional hyphen in RTF and Trados).

In preprocessing, we convert other fancy hyphens to - and strip other optional
ones.

The postprocess_plugin shows a long example or re-inserting non-breaking spaces
and hyphens into text, if that were to be desirable to the users.