Fix this autobuild suite for both ce_tmx.pl and tmx2lfl.pl and their Perl requirements.
OK, so ce_tmx.pl still needs to be called as "perl ce_tmx.pl", but tmx2lfl.pl
does not.  Later, we need to change ce_tmx.pl to use "#!/bin/env perl" instead
of "#!/usr/bin/perl", but that first requires getting rid of the -s option,
so it's a project for another day.