Speed up loading phrase tables.
I first thought the problem was inherent to strncpy(), because that was the
function that profiling said took all the time, and replacing it by strcpy()
fixed the problem, but it turns out the real problem is the n argument to
strncpy(): strncpy(dest,src,n) writes n bytes, regardless of the size of src,
because it pads with NULLs all the way to n.  So the real bug was setting n =
buffer size, instead of n = src size, as I should have done.

Nonetheless, this patch replaces strncpy() by strcpy() in various places, with
appropriate checks to make sure the operation is always safe.  strcpy() remains
at least slightly faster, and strncpy() remains a bad answer to a real problem.