Overwrite existing files by default.
joint2multi_cpt's -force option is now flagged as obsolete, and does nothing.
Existing files are overwritten by default, unless the file permissions prevent
the operation, in which case a fatal error happens, as would be expected,
when the program tries to open the file for writing.

It is not the job of these programs to refuse to do work requested because
a file exists - it is the job of the frameworks that call them to not call
them if they should not be called!