Committing the ibmcat test suite, created 3 years ago...
The ibmcat test suite checks whether ibmcat works correctly, with its default
arguments as well as with the -sort switch.

I put this test suite in my sandbox in 2009, but failed to add it, so now it's
finally making it into the repo, with changes made to the Portage R&D version
in 2010.