Large commit centered around putting the a= field in TPPTs
TPPTs, TextPTs, can now store the a= field.  This field is needed by the
unal features, some sparse features, and also for the future projects to
use the phrase-internal word-alignment from the a= field to do markup transfer
from source to target, both for truecasing (e.g., transfer sequences of all
caps) and for tags.

Other stuff:
- tppt.dump now outputs the textpt in LC_ALL=C sorted order.
- New Huffman coder, with binary format for use with memory-mapped IO.
- factor out the counting error callback in to errors.h/cc, since it is used by
  many suites.
- error.h now uses Current instead of dummy as the namespace for the current
  callback.
- .gitignore file has more stuff we can ignore
- New tp_alignment* series of file to store alignments in a memory-mapped file;
  their functionality is only partially needed for the a= field in TPPTs, but
  it's easier to bring them in completely instead of cherry picking the bits
  that are actually needed.