Add tokenize_plugin and sentsplit_plugin to PortageLive install list.
These plugins used to have to be left out by default, only included in relevant
systems, but the logic has been changed: for languages supported by
utokenize.pl, the plugins are never called, so it's OK to always install them
now.  In fact, they must always be installed, so that Chinese works out of the
box.