Add the -xtags switch to utokenize.pl to handle tags.
The tokenizer now tries to keep tags glued to the token they were attached to
before tokenization.  The process is error-prone, and I accept the fact that
some cases are not handled correctly.  On real test data, it makes a few
mistakes, but not too many.

Later revisions might try to go further with precision, but this is good enough
for a first tags handling baseline.

Cases that are handled well:
 - Tags on the left and right of a string are glued back to their neighbouring
   token.
 - Matched tags in the middle of tokens (i.e., the close and open are in the
   same token) don't break the token apart in most situation.
Weaknesses:
 - Recognition of matched tags in not done with a tag stack, only with with a
   linear search for an opening tag preceding a closing tag, or a closing tag
   following an opening tag.  Thus, <a><a></a> would incorrectly be considered
   to only have matched tags.
 - While punctuation at the end of the string is kept apart even when the tag
   handling routine says to keep a string together, punctuation inside a tag
   isn't handled right.  E.g., (a<a>b</a>c) would be tokenized as ( <a>b</a>c
   ), but a<a>.</a> would be tokenized as a<a>.</a> instead of a <a>.</a>,
   which would be preferable.