Fix test suites to be stable in parallel or not.
For a little while now, some test suites have been failing when run in
parallel, because of the PORTAGE_INTERNAL_CALL mechanics to suppress the
copyright notice.  Added PORTAGE_INTERNAL_CALL=1 in those Makefiles and
adjusted the references so the test suite output does not depend on this flag.

lm.train_mixture: this suite was failing because I forgot to add three files.
I forgot to add those files because the .gitignore entries applied both to the
files in the test suite and those in the ref/ subdirectory.  So I added
"!ref/*" and "!src/*" to the .gitignore files generated by ../Makefile.incl, so
that I don't ignore such files again by accident.