First pass detokenizing with tags.
The detokenizer in this commit always handles tags as such, performing
detokenization as if the tags were not there.

Unit testing has been written to make sure that the code detokenizes brackets
right even in the presence of this code (a bug in Perl 5.8 and 5.10, fixed in
5.14, makes this non trivial).

Unit testing for detokenizing with actual tags is not written yet.