Make hmm-lex unit test faster.
- Reduced from 5 to 3 iterations per model type (IBM1, HMM)
- Added filtering options -p 1e-3 -p2 1e-4

The filtering options help a bit with speed, a lot with the size of the
reference files.  That was necessary because reducing the number of iterations
blew up the size of the reference files, and this was the way to bring them
back to reasonable (although still larger than ideal).