Only one space when break a token because of tags.
When a token is broken up because it contains unbalanced or self-closing tags,
only one space should be introduced.

This commit fixes the known issue reported with detokenization in the previous
commit, because that problem was in handling the multiple spaces that the
tokenizer spuriously introduced, but not that we don't introduce them, we also
don't have to remove them latter.

Also, removed some obsolete code that was left commented out previously.