Add a test case that exercises case transfer from source.
The truecasing test suite only exercised a specific bug.  Now enhanced to also
exercise one situation of doing truecasing with transfer of casing information
from the source sentence.

This work is in preparation for regression testing changes to
markup_canoe_output, as well as for testing truecasing.pl, of course.