Improved intra-token tag transfer and whitespace output.
Move intra-token tags to start or end of token in output as appropriate.
If both tags are within the same token then move to the end of the
token; otherwise move tags to whichever end of the token there is a tag
at in the source.

Mimic the source whitespace more closely.

Added another intra-token tag case to whitespace input file.

Nested tags not handled yet.