Use real word-alignment info to transfer tags.
nbest2rescore.pl now writes word-alignment information in the PAL file, when
given the -wal switch (requires the -palout switch to have any effect).

markup_canoe_output now uses word-alignment info found in the PAL file, if any,
instead of the heuristic.  Heuristic alignment is still used when 1) the word
alingment info is missing, or 2) the -hwal switch is given on the command line.

For all this to work, markup_canoe_output now uses a vector<vector<Uint> > to
represent alingments, as we do elsewhere, isntead of the simplified vector<Uint>.

Various unit test cases were added to test this work.

Squashed commit of the following from my private markup-with-a-field branch:
2013-01-08 10:02:55 -0500 6a49987 Eric Joanis (markup-with-a-field) Pretty much finished WAL in PAL work.
2013-01-07 17:26:10 -0500 e902a5a Eric Joanis markup_canoe_output: read and use WAL info from PAL file.
2013-01-07 10:42:40 -0500 d1812b9 Eric Joanis parse_pal.cc now reads WAL info in PAL file.  -hwal switch to markup.
2013-01-04 16:24:31 -0500 74aeb1a Eric Joanis nbest2rescore.pl: output WAL inside PAL file.
2013-01-04 16:06:11 -0500 3fe4d10 Eric Joanis Commit before rolling back some work: WAL is still stand-alone here.
2013-01-04 14:32:56 -0500 93a0d9e Eric Joanis Add a -walout switch to nbest2rescore.pl and test it.
2013-01-04 11:55:08 -0500 634d079 Eric Joanis Regression test case for markup_canoe_output, plus -vv switch.
2013-01-03 13:01:08 -0500 326ecbc Eric Joanis Started to replace v<Uint> by v<v<Uint>>