Maintain proper nesting of empty tags.
Nested empty tags were previously pulled apart. For example,
<a><b><c/><b><a> became <a></a><b></b><c/>. Now, they remain nested.

Some attention has been given to the placement of point tags and empty
tags with respect to other closing/opening tags in the target text.
However, empty tags are still bound to the following token in the source
and are often placed with respect to that token in the target output.
This can lead to perhaps less than perfect results in the face of
cross-overs. To do otherwise would require the ability to bind to
another XML tag and an algorithm to decide which one (would also likely
require left and right binding options).

Added a fairly extensive emptytags test. The reference file matches the
current output. A ref/emptytags.P.tags.tok.ideal file was also added,
showing what perhaps would be more ideal results, but which are not yet
acheived, as per the discussion above.