Output OOVs with intra-token tag placement matching the source.
To support this we keep track of src token insertion points (lip,rip) in
the Elem structure for intra-token elements.

Empty elements at the start or end of a token that are nested within
intra-token elements were problematic because we did not recognize them
as intra-token during input, and they had incorrect lip and rip values.
Also such elements at the end of a token were associated with the next
token, so they were difficult to detect correctly and were at risk of
moving out of place even for OOVs.

This problem was resolved by running a second pass over the tags
(ordered by tag id) just after the initial input scan, fixing those
elements that need it based on information not available during the
initial pass.

This approach may also be helpful as a framework for solving other
ordering issues.

Added an oov unit test to the markup_canoe_output unit test suite.