Added some tricky cases to emptytags unit test and fixed whitespace.
Added tricky cases involving an emptytag in a sequence of right tags
where the output requires tag movement due to cross-overs.

We were incorrectly outputting whitespace to the right of a left tag
notmarked to have such whitespace in the situation where it followed
a point tag that was marked as having whitespace to its right. Fixed
by resetting the rwsr variable tracking this property.