Maintain original tag order for empty elements within tag sequence.
In the output, maintain the original order of tags when empty elements
are interspersed among non-empty tags in a sequence either before or
after a token.

This is accomplished by making a third pass over the tags, adjusting
the target token range for the empty tags in the sequence so that the
empty tags "move" with the non-empty tags in the sequence.

Updated reference for emptytags unit test to match new expected output.
Removed the ideal reference file for emptytags unit test as it is no
longer needed.