Treat OOV look-alikes as OOVs; make intra-token elements span tokens.
Updated code to recognize and treat OOV look-alikes as OOVs. An OOV
look-alike is where the source token matches one of the target tokens
in the assigned target phrase for the element. This applies only to
the transfer of intra-token elements.

Intra-token elements now span the source token they are associated
with to trigger proper assignment of the target phrase. As a
consequence, non-OOV intra-token elements can span multiple tokens
in the target.

To facilitate the transfer of intra-token elements, we now also update
the target range of non-OOV intra-token elements with than one target
token to move intra-token elements to the the first or last token of
the target phrase depending on whether the element's left tag is at
the start of the source token or not.

Added tests for OOV look-alikes to oov test file.

Fixed the whitespace and nested1 unit test target files to not trigger
looks-like-an-OOV. The tests contained some intra-token examples with
identical source and corresponding target tokens. Changed the case of
such target tokens so that they no longer "look like" OOVs. We leave
the OOV look-alike tests to the oov unit test file.