Added code to make point tags at start/end of a line stay there.
Empty elements preceding the first non-empty element or token at the
start of a line should remain at the start of the line. Ditto for those
at the end of a line, and regular or paired non-empty elements that span
a line. These last two cases were already true, but the first one
required some special handling to fix the target span for such empty
elements.

Added test cases for empty tags at start or end of a sentence.