Augment translate.pl references with P.txt.
Each TMX or sdlxliff test case now checks the XML output, as well as the P.txt
output, so that we can more easily inspect the result of doing the whole
pipeline with regular diffs.  This is particularly important for regression
testing and when making a small change, to make sure it has the desired
behaviour, changing only the lines where change is expected.

For now, several cases fail where I don't think the current output is correct.
This problem applies to P.txt and XML files .