ptable.encore-scores.cc now examines a larger header to determine the file format.
Instead of just the first 100 lines, we look at the first 10000.  The extra
memory required for doing so is not going to be significant, and it will mean
we'll notice the requirement for alignments even if they are scarce in the CPT.
In particular, this make the tm-count test suite pass.
(cherry picked from commit d98251e14045f3de198e8c000ceadd3c09fdc2eb)