Add XSLT formatted version of WSDL, for readably formatted rendering.
The wsdl-viewer.xsl file is Apache 2.0 licensed, so we can distribute it.
The .wsdl.xml file has the same contents as the .wsdl file, but adjusted for
rendering with wsdl-viewer.xsl.

TODO: automatically generate the .wsdl file from the .wsdl.xml file (or vice
versa) and install both when intalling a PortageLive server.
TODO: add the relevant third-party Copyright notice to our NOTICES file.