Add arabic sentence split and test case.
The Arabic sentence splitting test case is in unit-testing/tags.  Although a
misnamed location, this puts it in the same directory as other sentsplit_plugin
test cases that already exist.

Also reorganised the reference files in unit-testing/tags to place them all
under ref/, with subdirectories, instead of hidden in other directories.