Fix bug causing casemark.py to seg fault in translate.pl unit-test.
casemark.py was leaving a '\n' glued to the last token in a sentence
when creating the external vocab for the NC1 LM, so later if a lookup
was done on the token (without the '\n glued to it), it was not found in
the external vocab. The vocab access in LMtpt::mapId() assumed a valid
id (Voc::word() assumes a valid id) causing an out-of-bounds vector
access resulting in the seg fault.

Fixed casemark.py to correctly strip the '\n' when tokenizing the
sentence; it now calls the portage_utils.strip() function to tokenize
the input line.

Added an assert() in LMtpt::mapId() to validate the id to make this code
more resiliant.

Finally, while debugging this code, I observed that when querying the
NC1 LM, the start symbol (<s>) should be added to the context (but was
not) if it is less than LM_ORDER-1 in length because the NC1 LM was
created with start and end symbols added. This has been fixed.