Prepend all fatal error messages from Perl scripts with "Error: ".
This is to make it easier for the user to find errors in log files, e.g.
using grep.

Likewise, warning messages start with "Warning: ".

Always print an error message if exiting early with an error code.

Also, fixed some top comment blocks and/or help messages along the way.