PTGSH-193: translate.pl -nl=p or -nl=w fails in boscap.py
Fixed portage_utils.py split() function to handle empty or
whitespace-only string as a special case to fix the above
problem. split() was returning [u''] when it should have been
returning [].

Added a unit test in unit-tests/portage_utils.py to test the
empty/whitespace-only string case and in unit-tests/truecase.pl
to test the empty line case.

Also fixed an error message in boscap.py that was resulting in a
python exception in the case of mis-matching token counts with an
empty pal line.

Finally, tightened up the regular expression for tokenizing in
the presence of BOS markup in boscap.py and boscap-nosrc.py.