Factor utf8 clean up code from tmx-prepro and preprocess_plugin into clean-utf8-text.pl
We had mostly duplicated but not quite consistent rules for cleaning up utf8 in
tmx-prepre/template/preparation/Makefile and preprocess_plugin. As part of
doing PTGSH-117 (clean up NBSP everywhere), I factored out this code so we can
have just one place for that stuff.

NBSP handling: now all substituted by space.
Paragraph separator (PS) and line separator (LS): space

A bunch of other characters are handled too, and each is individually
documented in clean-utf8-text.pl, with the rationale for the handling choices
made, when it's not obvious.