Added -src-country option to translate.pl.
Also added an up front explicit check that the source locale is valid,
producing a meaningful error message to the user if not.

translate.pl now passes the source language country code to truecase.pl
via the -loc option.

The target language country code is not needed because most of the
truecasing code is locale agnostic, including boscap.py and
boscap-nosrc.py. (Python unicode string processing handles case checking
and conversions correctly without the locale set in all but the rarest
of circumstances.)