rescoring.metric test suite: alt references added; OK on MATS machines and GPSC.
The original references were from Balzac, with an old compiler and architecture, or
possibly even from before Balzac. The new alternate references are from the GPSC, but I
manually checked and they files generated on the MATS machines, which we know to be
reliable, are identical to those generated on the GPSC. So I declare these all good.