proposed GPSC optimization: psub -noc(luster) option.
Intended usage:
   psub -16 -noc -join make all

This command will run make all in a whole small node (16 cores, 96GB RAM), blocking
recursive job submission.

This can be useful when the job involves many small jobs launched by make. With a 30
second-per-GPSC job overhead, short jobs are costly. With -noc, the overhead is
incurred once and all the small jobs happen inside a single GPSC job.

I don't know how frequent this use case will be for others, but I'm constantly using it
when testing with toy data and writing the tutorial, so it's worth pushing at least for
this case.