PTGSH-316: Make BiLMs work with TPPTs.
My prototype BiLM implementation did not support TPPTs yet, because the
BiLMAnnotation was only created while loading a text phrase table.

Now, PhrasePairAnnotator::annotate has two variants, one that takes a
full PhraseTableEntry, one that takes just the source and target phrase.
Only the latter one has to be implemented, as a default is provided for the
former to call the latter, when that's appropriate. The two argument variant is
called when loading a text phrase table, the four argument variant is called
when reading a phrase pair from a TPPT.

Plus: toy test case in bilm/ test suite to exercise this new code.