PTGSH-376 Make the SOAP API backwards compatible again.
 - deprecated functions (getTranslation/2/CE, translateTMXCE, etc) are added
   back to the API and behave as they used to.
 - functions that had parameters added in the past (getTranslation/2/CE,
   translateTMXCE) now have appropriate default values if they are called with
   some parameters missing
 - getVersion() returns PortageII-3.0.1

Tested: a client that was programmed for the SOAP API from Portage 1.4.3,
PortageII 1.0, 2.[123] or 3.0 will work with the 3.0.1 PortageLiveAPI.php, even
if it was programmed from the WSDL from any one of those versions.

No clients still use older versions, so I'm not going back any further, but
1.4.1 and 1.4.2 had the same API as 1.4.3 so 3.0.1 is probably backwards
compatible with them too.