PTGSH-347: translate.pl: stop outputting meaningless "Empty input line" warnings.
canoe warns on empty input lines because that's usually a sign that something
is wrong with the input, especially when processing one-sentence-per-line data
files.

However, translate.pl uses empty lines to track paragraph boundaries (in -nl=p
and -nl=w modes) and is used to handle any user provided cut-and-paste
material, so empty lines are perfectly normals in translate.pl input.