PTGSH-386 train_tm_mixture: read and merge a= and c= fields during merging phase
Semantic implemented:
 - values of the c= field are tallied, if at least one table has it.
 - the most frequent value from the a= field is kept, which means different
   input tables are effectively voting.

The for a= field, in competitions we use a more complex logic, where we keep
the alignments from IBM4 ahead of HMM ahead of IBM2, but that is done using
external scripts via a separate process. For portage.framework and commercial
clients, we'll just keep it simple because it's not likely to make a
significant difference.