PTGSH-343: make libstdc++.so a copy instead of a symlink, to fix LD_PRELOAD error.
That error is caused by the symlink being broken when PortageLive is installed
via a CD or a Windows machine. We could add instructions to say don't do that,
but using copies makes the installation more robust, which is preferable.

For the same reason use copies instead of symlinks for secure/images and
secure/favicon.ico.