tm-count test suite: adapt to 4.9.2: a= and c= field can be output reversed.
When PhraseTable::write() uses tgt_p->second.annotations.write() to write the
annotations (a= and c=), there is not guarantee of what order they'll get
written in. Previously, we always had a= first and c= second, but that is not a
requirement, it was just by chance. Now let the references reflect that.

(cherry picked from commit f3b5f3aec649612a5f3325ee9ef4684a50a8eae2)