count_unal_words opt 9.7s=>8.4s GreenReader::operator() use char* buffer
By now, some 33% of the CPU time in count_unal_words is spent in strtok_r, so there isn't
much more I can squeeze out of this program. Even this optimization was not clearly worth
it, but I felt this relatively small gain for count_unal_words was worth implementing
because as a side effect it will speed up all code using the green alignment reader, which
includes parsing word alignments inside the decoder. That can only be a good thing!