Enable compiling with -DNDEBUG, where assertions are suppressed.
When assertions are suppressed, variables that are only used for asserts
trigger a warning as unused variables, which gets turned into an error since we
use -Werror. Add macro FOR_ASSERT(x) which adds a dummy use of x and avoids the
warning, simultaneously documenting that the variable exists only so we can
assert something about its value.

Also, code that assert(false) instead of returning a value now also triggers a
warning, so I added dummy return statements to such functions.