assert() statements should not have any side effects!
I know we don't use -DNDEBUG, but assert statements should still be doing just
assertions, not executing code that is needed for its side effects.

So I've reviewed all assertions in PORTAGEshared and re-written the code in the
15 cases I found where the argument of the assert statement is required for
correct execution of the code.

All the unit tests pass now, with -DNDEBUG, so hopefully I did not miss any...
But I know, we'll still continue to compile without -DNDEBUG by default.