Incr PTGSH-507 add whole sent pair if not phrase extract for that sent pair
We noticed that when a sentence pair generates no word alignment links, as can
be the case with OOVs on both sides, we might get nothing out of it, but in
incremental that's potentially bad because we know the sentence alignment is
highly reliable. Now we add the whole sentence pair as if it were one phrase
pair, when it generates no phrase pairs.