ULexitools.pm (de)tokenize_file(): close output file
We discovered a bug due to a race condition with downstream reader process: with
detokenize_file() and tokenize_file() not closing their output files, we depended on
buffer flushing to write the files out before the next process in translate.pl was reading
them, thus causing part of the output to be lost, in particular with -nl=p since
deparaline() gets called immediately after detokenize_file().

For some reason, this problem did not occur with the regular English-French pipeline or
many other we tested, but surfaced with the Arabic pipeline with hashtag handling. I'm not
sure why, but no matter, now the problem is fixed with close OUT; being called.