make g++ 7.3.0 happy by fixing lots of warnings
In particular, it is now a warning (and therefore an error with -Werror) to
have misleading indentation in the code.

E.g.:

   if foo
      bar;
      baz;

must be written

   if foo
      bar;
   baz;

Of course, the first example is terrible code formatting, but now it's also
disallowed!

Slightly more annoying is that compact code like this is no longer allowed
either:

   if (f_nbest) delete f_nbest; f_nbest = NULL;
   if (f_ffvals) delete f_ffvals; f_ffvals = NULL;
   if (f_sfvals) delete f_sfvals; f_sfvals = NULL;
   if (f_pal) delete f_pal; f_pal = NULL;
   if (f_wal) delete f_wal; f_wal = NULL;
   if (f_lattice) delete f_lattice; f_lattice = NULL;
   if (f_lattice_state) delete f_lattice_state; f_lattice_state = NULL;

It must instead be written like so, because again one might be mislead to think
f_n best = NULL is conditional, when in fact it is always executed:

   if (f_nbest) delete f_nbest;
   f_nbest = NULL;
   if (f_ffvals) delete f_ffvals;
   f_ffvals = NULL;
   ...

(cherry picked from Portage R&D commit 47f25b3f4f4827db000b6faa80a6f6f22a77b235)