SKIP tests that cannot run for missing optional dependencies, don't FAIL
When SRILM is not installed, three test suites cannot be run. Those are
not failures, really, they're just skipped.

Similarly, if generic-2.1 is not installed, even if
PORTAGE_GENERIC_MODEL is defined, that part of the nnjm test should be
SKIPPED, not FAILED.

Test suites should now exit with status 3 instead of 1 if they have to
be skipped.  ./run-all-tests.sh now interprets that 3 exit code
accordingly.